requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
newsapi-python==0.2.6
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.3
